refactor the plant helpers back to plant, but at the bottom
